refactor: abstract common functions from web search
